+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 ### Changed
 
 * Bump `async-utility` to 0.3 and `async-wsocket` to 0.11 ([Yuki Kishimoto])
+* nostr: remove self-tagging when building events ([Yuki Kishimoto])
 * lmdb: use `async-utility` to spawn blocking tasks ([Yuki Kishimoto])
 * pool: add `PingTracker` and improve relay ping management ([Yuki Kishimoto])